bwconncomp: return indices for object elements and not its boundaries.
* bwconncomp: PixelIdxList only gives indices for the borders of each
object in the input image. Fix this and give indices for each element
in the object. Also, by making use of bwlabeln, expand support for matrices
with arbitrary number of dimensions (it was only accepting connectivity
of 4). The default connectivity was changed to 8 for Matlab compatibility.
Add tests.
* private/make_conn.m: return second output argument giving connectivity
for the user (not used internally since for that is only the logical matrix).
* NEWS: make note of this big change. Also added fucntion to list of
functions supporting ND images.