Make the is* image type functions identify N dimensional images.
* isbw.m, isgray.m, isind.m, isrgb.m: return true for multi-dimensional
images where the 4th dimension can have any length. 3rd dimension must
still be size 1 for bw, gray, and ind, and size 3 for rgb. Add tests.
* isbw.m: remove the previously deprecated usage of isbw to check if all
elements were 1 or 0. Must be specified via the logic option. Return
true for MxNx1xK input. Skip comparison of the elements with the non-logical
option, if class is logical. Use != instead of == for small performance.
isind.
* isind.m: use isindex() directly since it's that's what should be optimized
to skip looking at the rest of the elements as soon as it finds a "bad" one.