imhist: bug fix for images with little contrast.
* imhist.m: when images have little contrast, i.e., only a small section
of the range of values is used, majority of the bins will have a value of
zero. The y axis are automatically readjusted to prevent high peaks from
hiding the histogram but when the median of the histogram was zero, it
failed since it tried to set it to [0 0]. Fixed by skipping automatic
adjust of axis in the case of such "bad" images.