connectivity: move the validate() static methods into the constructor.
* conndef.cc (Fiptconncheck): this function performs a bunch of logic
that is useful for a constructor from octave_value. But if we move it
there, then all the validate static methods are no longer needed, and
we can just check by calling the constructor.
(invalid_connectivity): an exception class so we can give meaningful
error messages on why the constructor failed (and the check failed).
(Fconncheck): make the position arg optional.
* conndef.h: declare new exception class for incorrect connectivity
arguments, remove the validate methods and add them to the constructor