Remove unnecessary get_padded_index function
Carn? realised that there was no need to loop over BW and then convert
the BW index into an L index. It was sufficient to loop over L, and
since the border for L is zero-padded, it will be skipped during the
computations. In particular, we will not have to worry about
out-of-bounds checks when considering neighbours, since we will never
consider the neighbours of L's boundary, which is all zero.

* bwlabeln.cc (get_padded_index): Remove this function. (bwlabel_nd):
  Loop over L instead of over BW. Remove the call to get_padded_index.