Improved Makefile and explicit includes.
* Makefile: set CXXFLAGS=" -std=c++0x" to all recipes rather than only one
place (we now need it for connectivity too). List %.h files as dependencies
so it gets rebuild when only the header changes, rename the target "test"
to "check" for similarlity to Octave core target names. 'addpath ("../inst")'
when testing to find @strel on the path.
* connectivity.cc, connectivity.h: explicity include string and vector
headers.  Remove default constructor since we are using C++11 now anyway.