Fix disk shape (broken since f1cf02800a87) and add option N periodic lines.
* @strel/strel.m: the change to fspecial with cset f1cf02800a87 change the
size of the disk shape.  This change uses the code previously on fspecial
to get back the disk shape.  In addition, we also add the option N for the
disk shape but require it to have a value of zero.  This is because our
method is equivalent to Matlab's N of zero (no approximation by periodic
lines), but Matlab default is 4 and we do not want to introduce backwards
incompatible changes when we do implement the option.
* NEWS: make note of this change.