Fix package dependencies and fix some workarounds for older Octave releases.
* DESCRIPTION: we have been dependent on the next Octave release for a while now
so we might as well set dependencies to 4.0.  This should force anyone installing
development version to use -nodeps and be aware of possible problems.  Also drop
dependency on the general package since Octave 4.0 will have inputParser (we were
only dependent for @inputParser).
* imcomplement.m: the fix to bitcmp is now on 4.0. We will already be dependent
on the new version so let's use bitcmp() for performance.
* imattributes.m: drop notice about release 4.2 (which is now 4.0)
* montage.m: adjust notice about missing newer Octave versions. We should keep
this check until it's unlikely to have the general package with @inputParser