maint: have Makefile use target/ when creating packages and documentation.
* Makefile: when using th epackage dist, install, or release targets, the
tarballs would be dropped at the root of the package.  This makes use of a
target/ directory where all this things are dropped.  This is cleaner but
its main advantage is easier to rule on hgignore.  The fact that the
filenames had version numbers on it, also made it worse.
* .hgignore: ignore everything under target.