watershed: fix bug when flooding watershed lines.
* src/watershed.cc: when the flooding was moving into watershed lines, it
was crossing it because it kept the label of where it came from.  Fix this
by not keeping the label of the element that added it to the priority
queue.  This should also reduce memory usage since we don't keep the label
value for each pixel in the queue.  This makes the function pass more of
the failing tests (but it is still not full Matlab compatible).