Fix build when CXX compiler does not need extra flags to support C++11.
* src/configure.ac: we are assuming that CXX does not support C++11 by
default and that OCTAVE_IMAGE_PROG_CXX_CXX11 will either error or add
flags.  If there is no need to add a flag, then CXX will not be modified,
and the sed code to remove the CXX would fail (since we expected a space
after CXX and before the flag).  The end result is that CXX would be
appended to MKOCTFILE leading it to be something like "mkoctfile g++".
Simply remove the space, there is no problem if in the end there is an
extra space in the command.
Note that development versions of Octave will enable C++11 support by
default by adding any required flag to CXX (autoconf treats these as part
of the compiler, not CXXFLAGS).  Point is, even with those flags, problem
remains since it's in CXX from the very start.