imerode/imdilate: fix for an empty ([]) SE.
* imerode.cc: remove special code path for an empty SE. We want the erosion
functions to behave correctly when nhood has no elements (actually, current
behaviour is incorrect since it returns the input but should be all Inf).
* strel.cc: behind the scenes we just convert an empty nhood into a SE
with all zeros (scalar).  This allows origin to be computed correctly
(what's the index center of an empty array? (0,0) are the coordinates
in C++ of the first element), which means that pad_array will return
something instead of [-1, -1].
* NEWS: make note that were small changes but don't bother with details
since an empty SE is not important enough.