maint: set CXX when running configure from maintainer Makefile.
* Makefile: when pkg installs a package, it sets CXX via "mkoctfile -p CXX".
Do the same in our maintainer Makefile so that  "make all" from the repository
behaves the same as when installing the package.
* src/configure.ac: put check for octave and mkoctfile at the top, next to
the check for sed, and the check for CXX after.  Shouldn't make any difference
at the moment, but we might need soon to get mkoctfile first to query for
what CXX to use.