Overhaul error handling in oct files (bug #57585)
With Octave 4.2 the global error_state variable was deprecated and the
error and print_usage functions started to throw exceptions which are
then caught by the interpreter.  Many of the oct functions in the
image package predate this and did not kept up to date but with the
upcoming Octave version 6, this throws warnings at build time.  This
commit touches over many related things, but can be summarised:

1) print_usage and error functions now raise an exception causing the
C++ function to return.  The 'return octave_value_list ()' after them
is never reached and no longer necessary.  This commit removes those
extraneous returns.

2) check of error_state was previously required after calling the
octave_value methods that get/convert to a specific type, to ensure
that the converted was valid.  The error_state is now always zero and
checking it was doing nothing, and the octave_value methods are
raising an exception instead of setting error_state.  This commits
removes the check of error_state and lets the exception through.  This
means less check but also that the error message will be less useful
since it won't be clear to the user which variable caused the failure.

3) the connectivity.cc file had most of the changes.  Exceptions and
checking for error_state was happening both inside the connectivity
class and Fconndef function which made things messier to change.  This
was mainly happening because of the connectivity constructor that
takes an octave_value and that was required for use in other C++ code.
This commit removes this constructor and instead adds a "normal" C++
function, same as Fconndef, moving the logic of octave_value out of
the connectivity class.

4) like the connectivity class, the strel class should lose the
handling of octave_value.  However, it is two tighly integrated with
the octave @strel class.